refactor: 메시지 상수 파일 통합
