--- src/page/home/constant/endSection.ts
@@ -1,1 +1,0 @@
-export const END_MESSAGE = '지금 바로<br/>NiNEDOT와 함께 하세요!';
