--- src/page/home/constant/startSection.ts
@@ -1,3 +1,0 @@
-export const TITLE_MESSAGE = '목표를 세운 순간부터\n시작된 변화';
-export const CONTENT_MESSAGE =
-  '계획만으로 끝나지 않도록,\n실천까지 함께하는 만다라트 루틴을 경험해보세요.';
